\name{autonorm-package}
\alias{autonorm-package}
\alias{autonorm}
\docType{package}
\title{
\packageTitle{autonorm}
}
\description{
\packageDescription{autonorm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{autonorm}
\packageIndices{autonorm}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{autonorm}

Maintainer: \packageMaintainer{autonorm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
